Restructuring, improved installation & added microblog.rss
If you ask for a file microblog.rss, we now give you the RSS feed of the first (and likely only) admin user.

All actions and formats are now in separate files, making it easy to add new formats (like JSON?) and new actions (they can even be implemented in terms of each other, like the microblog action).

When installation hasn't happened yet, we automatically bring up the "New Admin user" screen now.

Improved docs so people have an idea how this is intended to be used.